ola: workaround aclocal preferences
OpenWrt autoconf is configured to prefer global aclocal macros over
project specific ones. This is not ideal as OLA ships with its own
version of ACX_PTHREAD that adds support for C++. Autoconf-archive
provides AX_PTHREAD and alias for it ACX_PTHREAD but that handles C.
Using that version results in to undefined CXX variable and thus broken
build.

Better fix would be to prefer local scripts over the global ones as that
is how it works in other distributions. This is just a workaround to fix
compilation.

This renames macro and adds alias for correct ACX_PTHREAD. This
effectively forces autoconf to choose local version as there is no
ACXX_PTHREAD defined in global aclocal directory.